requirements.txt
django==4.0.3
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
drf-yasg==1.20.0
factory-boy==3.2.0
